embedder/requirements.txt
open_clip_torch
pillow
torch
torchvision
